LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-normalization/normalize-image-expected.txt


PASS Normalizing a valid <url> returns a CSSImageValue
PASS Normalizing a bad <url> returns a CSSImageValue
FAIL Normalizing a <gradient> returns a CSSImageValue assert_equals: expected "CSSImageValue" but got "CSSStyleValue"
